requirements.txt
langgraph
langgraph-cli

langchain
langchain-google-genai
python-dotenv
psycopg[binary]
pytest
numpy
pydantic
